Updated many Unit functions and procedures
Added SizeOfString to Functions.
Fixed Mode-changing procedure for Screen.
Added new Text procedures:
PrintInv - Prints text to screen in Inverse
Updated Print to output in non-inverse
Changed comment for PrintNumber
Added PrintNumberIntoNoCR to print an Integer to the screen without a Carriage Return appended
Same for PrintNumberByteNoCR
And PrintInvNoCR
These are so you can print on the bottom line of the screen without it automatically scrolling up one line.